Complete project setup and Go module initialization
- Add go.mod with module path and dependencies (amqp091-go, zerolog)
- Create directory structure per DEVELOPMENT_PLAN.md
- Add .gitignore for Go projects
- Add Makefile with build, test, clean, and docker targets
- Move dataset files to datasets/ directory
- Fix uint16 type cast for ChannelMax in broker/rabbitmq.go

Closes #1

+++ internal/broker/rabbitmq.go
@@ -56,7 +56,7 @@ func NewRabbitMQ(ctx context.Context, cfg RabbitMQConfig) (*RabbitMQ, error) {
 		},
 	}
 	if cfg.ChannelMax > 0 {
-		config.ChannelMax = cfg.ChannelMax
+		config.ChannelMax = uint16(cfg.ChannelMax)
 	}
 	if cfg.FrameSize > 0 {
 		config.FrameSize = cfg.FrameSize
